requirements.txt
Flask
websocket-client